Rename `bytes.rs` to `bytes_generator.rs`
