+++ src/compaction_execute_bench.rs
@@ -14,7 +14,7 @@ use tokio::task::JoinHandle;
 use tracing::{error, info};
 use ulid::Ulid;
 
-use crate::bytes::OrderedBytesGenerator;
+use crate::bytes_generator::OrderedBytesGenerator;
 use crate::compactor::WorkerToOrchestratorMsg;
 use crate::compactor_executor::{CompactionExecutor, CompactionJob, TokioCompactionExecutor};
 use crate::compactor_state::{Compaction, SourceId};
